Revert "Remove yMargin from text editor scroll"
This reverts commit 2b83008ad642e369f60ad1dc71bd21a4a6041202.


Revert "Fix jumping to a wrong location after tab change"

This reverts commit 4c25e2f022864cdc91e50eec6f9befaffe6072cc.


Revert "Fix edit breakpoint dialog causing wrong scroll"

This reverts commit 1f637d76af956d5150e15a16c2f9451a5e744db8.

Reason for revert: This change has caused a series of
issues:
* crbug/1368113 (Then fixed in crrev/c/3921602)
* crbug/1370955 (Then fixed in crrev/c/3932607)
* crbug/1370991 (Not fixed yet)
and the initial CL was already in the branch of M107.

So, in this CL, I'm reverting all the chain of changes
and back-merge it to the branch. I'll back-merge this to
the M107 so that the initial fix causing crbug/1368113
won't be in M107 as well.

I'll tackle the above issues later at once since the initial
bug wasn't critical.

Original change's description:
> Remove yMargin from text editor scroll
>
> * Fix going to lineNumber + 2 instead of
> lineNumber + 1 (textEditor.toOffset already
> adds +1 to account for the index difference)
> * `scrollIntoView` adds 5px default margin
> from the scroll position so the scrolled
> line is somewhere between
> correctLine and correctLine + 1 (Since
> we were scrolling to correctLine + 1 with 5px
> margin)
>
> This fix updates the behavior to:
> * Jump to the start of the lineToScrollTo
>
> Fixed: 1358205
> Change-Id: Ib21c82b922ab1663276c7f33b576a15bf01743f6
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3892774
> Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
> Commit-Queue: Changhao Han <changhaohan@chromium.org>
> Reviewed-by: Changhao Han <changhaohan@chromium.org>

Change-Id: I9ed7f40df71ab19b65d271fcf68be704bba346b9
Bug: 1370991,1370955,1368113,1358205,1371435