Use static margins while deciding on where to restore scroll
Trying to get offset information using `coordsAtPos` sometimes
returned `null` even after the view is rendered, especially on Mac.
This caused tests to be flaky (sometimes we're adding margins sometimes
we're not adding them)

The margins between the first line and the top of the content
and the left of the content and the first character is defined statically
in CodeMirror (4px from the left, 4px from the top). So, instead of
trying to extract this information from the editor itself, we embed
this to the TextEditor.

Bug: 1368113,1370991
Change-Id: I0af8a653ed7fc6ba39dcbf735b26711c7bca61a9