Remove animated scroll on breadcrumbs auto-scroll
When the breadcrumbs needed to scroll a crumb into view, they were doing
so by smoothly scrolling. However, if the user has loaded DevTools and
selected a deeply nested element, they end up with a really slow scroll
which is distracting and not useful.

This CL updates the behaviour so we scroll smoothly if the user clicks
the buttons manually to scroll, but in the event we need to adjust in
order to reveal the selected element, we will just jump right to it.

Fixed:1369734
Change-Id: I20738eec239a795aba5cefa197c8ca17a3f6ef99