[debugger] Further speed up call stack function lookup.
Another follow up fix for https://crrev.com/c/3720713 and
https://crrev.com/c/3934889 which avoids loading the script content and
computing the line endings for all scripts that have no source map or no
named entry for the start location of the function scope.

This speeds up the construction of the Call Stack items for the repro in
https://crbug.com/1369002#c6 by another 10x and is mostly
indistinguishable from M104 now.

Bug: chromium:1369002, chromium:1339683, chromium:424201
Change-Id: I5daf95c0d1e88189edbaf9e621d86ce7e33253c7