Create patch set 2
Uploaded patch set 2.

Replace checks for non-null parentTarget with a check for a type == Frame in the NetworkLogView.

Patch-set: 2
Patch-set-description: Replace checks for non-null parentTarget with a check for a type == Frame in the NetworkLogView.
Subject: Replace checks for non-null parentTarget with a check for a type == Frame in the NetworkLogView.
Commit: 89993f7075700345539046a04e49c46106de51e5
Tag: autogenerated:gerrit:newWipPatchSet
Groups: 89993f7075700345539046a04e49c46106de51e5