Parse status text from response headers text
In http://crrev.com/908142 I made the statusText field in
responseReceived no longer accurate. At first I also added a new
statusText in responseReceivedExtraInfo to account for this, but later
removed it in the code review due to this comment:
https://chromium-review.googlesource.com/c/chromium/src/+/2856099/37..41/third_party/blink/public/devtools_protocol/browser_protocol.pdl#b5950

In that thread I agreed to parse the statusText in the frontend out of
responseReceivedExtraInfo's headersText, but I forgot to actually do it.

This patch parses the status text from the response headers text by
following the same logic in net::HttpResponseHeaders::GetStatusText:
https://source.chromium.org/chromium/chromium/src/+/main:net/http/http_response_headers.cc;l=543-561;drc=e27059d25adea5e63c37974fa0633f58f3fd4f3b

Fixed: 1351286
Change-Id: I634b347cd5e8fdd5a591d3e7bf1d29fbc1c474c8