Fix TextEditor cuts off from the first line after refresh
Includes fixes after introducing initial fix:
* Jumping to wrong source location when tabs
changed in sources panel
* Jumping to the start of EditBreakpoint dialog
when it is opened.

For the tests, instead of creating an E2E test
for testing the behaviors from a high level (which
were flaky), created two unit tests that tests
the behavior of TextEditor.

Introduced a helper `ManualPromise` for being
able to wait for events/callbacks.

Change-Id: Iaa368c23ca73c98ab96b8b20f9122ad26ebe2a6a
Fixed:1368113,1370955,1358205