Strike through overloaded lengths as well
Even though the container had `text-decoration: line-through` applied,
it wasn't shown over this component. The reason was, `text-decoration`
does not apply to `inline-block` elements[1].

There were two options to fix the issue:
* Update `CSSLength` to be `display: inline` which caused unnecessary
white-space between separator, component and the `;` text. Since
it also contains a `block` level like element inside, I didn't want
to make this an `inline` element as well.
* Pass explicit props to `CSSLength` and handle text-decoration inside.
This is the solution I'm going with.

[1]: https://www.w3.org/TR/css-text-decor-3/#line-decoration

Fixed: 1374357
Change-Id: I4a5a965f719d96427789ba899d3379d97148754e