[CodeMirror] Restore CodeMirror5 like syntax highlighting.
In particular with CodeMirror5 we used to have appropriate highlighting
for definitions, which was lost when we switched to CodeMirror6. With
CodeMirror6 the `defaultHighlightStyle` (in `src/highlights.ts`) already
gets us the proper highlighting of definitions, so we can just take
inspiration from that.

Due to the way that the tags are implemented in CodeMirror6, we do
however get this at the expense of losing the 'token-variable'
annotation right now (which doesn't matter, since we didn't style that
meaningfully anyways).

Before: https://imgur.com/NmXA4K1.png
After: https://imgur.com/I1bNBBS.png
Fixed: chromium:1290182
Change-Id: I0c938df7459d1dccb6e8f35535aa15f607116ee9