[BreakpointsSidebarPane] Use Lit's repeat directive.
Our use of Array#map() to render the lists of breakpoint groups and
items doesn't work correctly for HTML elements that have state (like
<input type=checkbox> in our case). See the relevant parts of the
documentation for details[1].

This change introduces an `id` for `BreakpointItem`, which is basically
the UILocation's id, and uses the `url` as key to unique identify the
rendered `BreakpointGroup`s.

Drive-by-fix: Remove some dangerous uses of `as` and use `Promise.all`
in `BreakpointsSidebarController`.

[1]: https://lit.dev/docs/templates/lists/#when-to-use-map-or-repeat

Fixed: chromium:1375294, chromium:1375226
Change-Id: I09e17386ee02d875aefe756bb4b114a8832f7fab