[CodeMirror] Restore CodeMirror5 like syntax highlighting.
In particular with CodeMirror5 we used to have appropriate
highlighting for definitions, which was lost when we switched
to CodeMirror6. With CodeMirror6 the `defaultHighlightStyle`
(in `src/highlights.ts`) already gets us the proper
highlighting of definitions, so we can just take inspiration
from that.

There seems to be a subtle bug somewhere in CodeMirror6 and/or
@lezer/highlighter, which prevents the `token-definition` from
being picked up for function names. So that still remains to
be investigated.

Before: https://imgur.com/NmXA4K1.png
After: https://imgur.com/491h0Cw
Fixed: chromium:1290182
Change-Id: I0c938df7459d1dccb6e8f35535aa15f607116ee9