Fix hovered row colors for data grid in forced-colors mode
This is a follow-up to
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3753276

The change was intended to remove striping behavior for data grid rows
in high contrast mode to improve readability, and to indicate selected/
hovered rows with a highlight color.

The above change added a ":not:hover" specifier to the CSS selector
covering background color for odd-numbered data grid rows. Although
hover colors were working in high contrast, the ":not:hover" addition
meant that no rule was covering hovered elements in normal color mode,
so odd rows would flash unstyled if hovered.

Selected rows were also not quite working in high contrast. The "canvas"
rule was taking precedence over the "highlight" rule for selected rows.

This reverts the original change and adds the extra
":not(.dirty):not(.selected)" specifiers to the forced-color rules so
that they will have the correct precedence w.r.t the to normal rules
and each other.

Data grid row colors now work as expected in both normal and high
contrast color modes. There is no special hovering behavior in normal
mode, and both hovered and selected rows are highlighted in high
contrast mode.

Bug: None
Change-Id: If0206ca79f07f8542f1ff71bf52de00af9b48d98