[i18n] Fallback to default locale (en-US) when using wrong placeholders
When UI messages update placeholder names then the i18n library can get
confused since translations other than en-US and en-XL won't get
updated. This means that at runtime we'll pass the wrong placeholder
name e.g. for the de-DE locale.

This CL builds in a fallback for those cases. Rather then throwing
we show the message in the default locale which is guaranteed to have
the right placeholders as we check that via linter.

Note that to make this work we now have to memorize the locale for
the LocalizedStringSet cache, otherwise we'd provide the
de-DE strings even though the en-US were requested. This is
because the initial code was written with the assumption that
the locale will never change per DevTools run.

R=alexrudenko@chromium.org

Bug: 1376012
Change-Id: I89c73d294c33acfa616da3415c1138e38f2ac0ac