[sources] Preserve line endings when editing files
Fix bug where V8 considers every line changed when live editing a
script. The reason is that we "forget" about line endings and always
call `Debugger.setScriptSource` with '\n' instead of the original line
endings.

CodeMirror actually knows about the line endings, so it's a matter of
retrieving the current CM content with the right API.

R=jarin@chromium.org

Fixed: 1362193
Change-Id: Ia31cd0917d289b629304fc295a1f2a660307df20