Don't log CDP errors to console in DOMModel and DebuggerModel.
This is causing noise in unit tests and isn't useful outside of tests.
Overall this doesn't seem to be a right place to log errors, we might
want to think about a consistent approach to propagating errors, but for
now I consider logging to be no different from ignoring the error.

Bug: 1376495
Change-Id: I4aa770c9c4455a2325e44220111192e16ed2b1a0