[mac] Emacs movement keyboard shortcuts in autocomplete.
This enables the use of emacs movement keyboard shortcuts (Ctrl-n and
Ctrl-p) for the autocomplete menus in CodeMirror, similar to how we
already support these for the Console history. Note that this is only
available on Mac, since the browser doesn't forward Ctrl-n to the
renderers on other platforms.

Fixed: chromium:1167965
Change-Id: I45ce48fafd4cd7ff831b5d17a2759f3963230251