Header overrides editor: fix updating header overrides
Previously, when editing either a header value or a header name from
'foo' to 'bar', an event would be generated correctly. But on another
edit which reverts the previous one (from 'bar' back to 'foo'), no 2nd
event would be emitted. This CL ensures that a 2nd event is emitted in
such cases.

Bug: 1297533
Change-Id: If5028b7dedaa33e18d518290e1d06283a67c3f8f