[DevTools] Use function info from plugin in Performance panel
Whenever a plugin (f.e. DWARF extension) is available, try to resolve a
function's name by requesting the plugin for the information
corresponding to that function.

Bug: 1364601


[Performance] Translate function names in the timeline flame chart

The implementation is analogous to that of [0]. However, since the
source of the call frame data is different, (Debugger domain vs Sampling
data) extra steps have to be taken to extract the script's information.

In particular the 'target' or the context of the sample has to be
provided, once the target is availble, the corresponding script and
source map can be queried.

[0] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3720713

Bug: 1364601


Make PerformanceModel async

In preparation to resolve function names in the perfomance panel using
source maps, this CL implements two functional changes:

1. The loadingComplete function of the PerformanceModel is set to be
   async. This is because the name resolving will require the result
   of async invocations and this function is ideal for adding the
   calls into (since this function commands the trace processing after
   all trace events are available). With this change, other functions in
   the call chain are updated accordingly (this includes calls in the
   legacy TimlineTestRunner)

2. To avoid race conditions, the PerformanceModel instance owned by the
   TimelinePanel is not used until the TimelineController has finished
   the processing pipeline. Before this CL, this waiting didn't happen
   made the code subject to race conditions, which were made evident
   with the implementation of (1).

Tests in Chromium land that make use of the updated helpers in the test
runner need to be updated before this CL can land.

Bug: 1364601
Change-Id: Ibb834d36a4d3af020092358dcf202bd4780e194d