[Sources] QuickOpenOutline powered by CodeMirror 6.
Initial version of QuickOpenOutline using the CodeMirror syntax tree to
construct the outline for the Ctrl+Shift+P / Cmd+Shift+P quick open.
This doesn't add any new features (except the support for CSS in HTML)
and aims to be 100% compatible with the previous approach.

Doc: https://goo.gle/devtools-codemirror-outline-design
Bug: chromium:1377747, chromium:1087762
Bug: chromium:1167083, chromium:1167084
Change-Id: I25458d523c3110423136271ffb3d4adee80ea3f4