Resolves Bug: 1379182 and supports high contrast light and dark themes
This patch ensures the `-theme-with-dark-background` class is only applied when Devtools is in regular Dark Mode or High Contrast Dark Mode.
Some forced-active media styles weren't applied after toggling a light HC theme from a Dark Devtools theme.

This was because in HC, there needed to be a corresponding css rule with same  or higher level of specifity as the css rule in regular Devools Dark theme state (i.e. rules containing -theme-with-dark-background class).
The patch takes out this requirement and ensures we revert back to the user's Devtool's theme preference in Devtools after toggling back and forth between HC themes and Devtools Themes.


Bug: 1379182 - Console message links are disappearing in HC light themes
Change-Id: I8126469bab04cb6b5d2d360ccbd37d4e1cc4a80c