[Sources] Properly support TypeScript in the QuickOpen outline.
Most of the TypeScript support we already get for free from the basic
(permissive) JavaScript support. The main difference here is that we
also mark abstract methods as such, since they don't contain code and
might thus be irrelevant for the purpose of debugging; therefore making
it easy for developers to immediately see that is likely important.

Drive-by-fix: Also add support for array and object patterns in
parameter lists.

Doc: https://goo.gle/devtools-codemirror-outline-design
Fixed: chromium:10877762
Bug: chromium:1377747
Change-Id: If0d7a92637529a471f76b2f5210bc4dbcbda9253