Add label to elements style filter
Adds a new label to the styles filter on the element page to indicate
how many style rules are hidden by a given filter.

The new label and the existing 'no matches' label are now both marked
as labels so screen readers will read them out as they change as a
result of the input, making the experience for those users much better.

Bug: 1379450
Change-Id: I56f215275759b7604b4e1471b9e8a3695951ea40