Show popover when hover on CSS variable
Currently, there will show the CSS variable value on hover as a title. But, the time that waiting for the title is a bit too long, many people even don't know that they could get the value by hover.
Also, although you can jump to the definition to get the value by clicking on it, it is still very inconvenient.

This commit use popover instead of title to show the CSS variable value on hover. On top of that, we can extend more functions later.

Previously when the cursor was located outside the variables of an expression, the value of the entire expression would display.
The following commit highlight the part of value in effect，so tooltip on whole expression has been removed.
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3983512

Video: https://imgur.com/a/aSGAvNJ

Bug: None