Fix can't hide popover when scrolling
In the Styles panel, we have a number of popovers that need to be hidden when scrolling.

ColorSwatch related popovers already handle scrolling events separately. This CL unifies the handling of missing popovers, such as `hintPopoverHelper`, `imagePreviewPopover`. There is no need to handle scroll events separately when adding new popovers in the future.

Bug: None
Change-Id: I3ceb9c5157add6c245c71ce5e5a09f890c00e85a