Store `set-cookie` headers during request interception
During request interception there are 2 sets of `set-cookie` headers:
the ones from the original response as received over the wire, and the
ones from the header overrides. This CL merges both sets and persists
them on the `NetworkRequest`, because a consequence of request
interception is that no `set-cookie` headers are present in
`NetworkRequest.responseHeaders`.

Later on, when displaying all headers for a request, DevTools will
then be able to make use of the persisted `set-cookie` headers.

Bug: 1379319
Change-Id: I79b8a499a090f3515b361a46e74cf36ddde383d7