[BreakpointsView] Update keyboard navigation
This CL updates the keyboard navigation to unify the keyboard
navigation of the whole BreakpointsView. After this change, we only
have one focusable element throughout the view, and can use all
keyboard navigation (arrows, Home, End) on the entire view, i.e.
including the checkboxes for pausing on exceptions.

Fixed: 1379786
Change-Id: I6eb57a92fc7c34c26fdef96136bf03645fcd042c