[BreakpointsView] Make keyboard navigation more straight-forward
Previously we were using a map from a domNode to an element id in order
to be able to remember a selected element. Whenever the focus was
shifted from one element to another element, we needed to keep track
of the selected element, _and_ re-render the whole view to update
the tabindices. Sometimes, this re-rendering would also cause some
side-effects.

Instead, keep the information on which element is focused only in
the dom, and update when needed using coordinator.write(). This
directly updates the affected elements without needing to re-render.

Bug: 1346231
Change-Id: I7512c57eed95106b84b002ce35958371ff47cf05