[BreakpointsView] Fix auto-expand breakpoint group
If a breakpoint group was collapsed, adding a new breakpoint should
automatically expand it. This was only flakily working before.
Previously we correctly set the data on the BreakpointView to expand
the group, but this information was not used to re-render the view.

By making use of the live directive, changes in the expanded property
of the breakpoint group should also trigger a re-render of the
details element.

Drive-by: Make settings global.

Fixed: 1379813
Change-Id: I5f19b93589c425d58cec0724e70a620674a3d188