[console] Improve automatic wrapping of object literals.
In the Console REPL we try to detect object literals and wrap them into
parens, as otherwise the JavaScript engine would interpret things like
`{x: 2}` as block statements with a label `x` and an expression
statement `2` inside. But for developers it's much more useful to have
the ability to write `{x: 2}` to construct object literals quickly,
without having to explicitly wrap them properly as `({x: 2})`.

Previously the detection was quite sloppy and was basically just
checking whether the string starts with { and ends with }. But that also
led to `{}+{}` and other similar constructs being detected as object
literals and being wrapped improperly. While these are arguably (and
fortunately) edge cases introduced with ES2015+, it's still worth paying
attention here and reducing surprises for developers.

So with this change we explicitly check for trailing semicolons (as
called out in https://crbug.com/1246866#c14) and we also check that the
leading { is only closed by the final }.

Before: https://imgur.com/MeJJ8kA.png
After: https://imgur.com/1Sfjohu.png
Fixed: chromium:1246866
Change-Id: Iedb43c5a0e46f36368db7c45a1314ed4d844d4ce