Hide soft context menu on DevTools resize
It seems I managed to land the original "fix" (crrev.com/c/3532901) with
the crucial lines that hid the menu commented out.

This CL ensures that when the DevTools window is resized, any soft
context menu is hidden (because otherwise it is shown on screen in a now
incorrect position).

Fixed: 1378934
Change-Id: I84b6a43c1f141f2e11c939673346f8b9cef3ab29