Trash bin icon for header overrides editors
Use trash bin icon for removing header overrides and adapt white space.

(DevTools already has 2 other trash bin icons, one is coloured solidly and the other is used in the recorder panel. This CL adds a third one,
taken from Material UI icons. We plan to consolidate all icons soon,
but this is out of scope for this CL)

Screenshots:
https://i.imgur.com/vULR947.png
https://i.imgur.com/oJ1h2ld.png

Bug: 1297533
Change-Id: I48bc336e4407800398129034c7bbe3f5a602d09d