[sources] Remove unnecessary JavaScriptCompilerPlugin.
The JavaScriptCompilerPlugin extends the 'Sources' panel with a "live
compile" feature to perform an early check for syntax errors. This seems
generally useful theoretically, however in practice, this cause more
harm that adding usefulness today. The primary reason is that the plugin
takes the source code out of context and tries to compile it as a
script. This worked well in the pre-ES2015 world, but starting with
modules, this already didn't work (and was simply disabled for ESMs at
the time). With snippets being run in REPL mode (and allowing toplevel
await there), we have another counter example.

Generally, it seems much better to rely on just attributing the errors
from actually executing the snippet or errors from LiveEdit upon saving,
and that's what already happens today.

Therefore we are removing the JavaScriptCompilerPlugin, since it doesn't
add value on top of surfacing the actual execution / liveedit errors.

Fixed: chromium:1083344
Change-Id: I102f7a897b74f9e61d8d254f8b2f8a425d300b24