[BreakpointSidebarPane] Use the breakpoint storage id as item id
This CL makes use of the breakpoint storage id to uniquely
identify breakpoint items in the BreakpointView. These ids are
used for stable DOM, i.e. when re-setting the data on a BreakpointView,
LitHtml will make sure that states (such as the checkbox checked
status) remain stable.

Before, we used an id that was based on the UILocation. The problem
here is that the same breakpoint may have different UILocations
depending on its enabled state. The same breakpoint may therefore
be treated as different, as it was identified by its ui location.

This caused the following bug:

1. a breakpoint was set on a comment and slid to the next breakable
   line
2. the user deactivated the breakpoint
3. the UI location changed back to the comment
4. although the breakpoint itself is the same (just that the
   (UILocation has changed), the BreakpointView treated them
   as separate
5. the breakpoint item line should have been focused after
   clicking on it, but it wasn't since it was treated as a
   different item within the BreakpointView. I.e. the UI
   removed the DOM element that belonged to the previous
   UILocation (which was focused), and replaced it with a new
   (unfocused) one.

When using the breakpoint storage id, changes in the UILocation
no longer affect the focused state.

Fixed: 1381041
Change-Id: I41edcf13886a19a6778b4657be0b96cb37a02817