Add pencil buttons to initiate header override editing
Add a small button with a pencil icon next to each response header on
hover. Clicking this button does all of the following if necessary:
 - Prompt the user for a folder to store overrides in
 - Turn on the overrides setting
 - Create a '.headers' file to store header overrides in

The user ends up in the same panel which shows the response headers,
only that they are now editable to create header overrides.

Video: https://i.imgur.com/nSqcwLj.mp4

Bug: 1297533
Change-Id: I974b3314b504ef69d041495f291dcdee39c4db4d