Navigate to about:blank before reload & record trace.
This CL aims to solve the issue where a user hits "Reload & record", but
when they see the resulting timeline there are screenshots from the
previous page included.

There are two ways we could solve this:

1. Keep the reloading logic the same, but try to filter out all events
   that occured before the first navigation. This works easily for the
   simple case of "Reload & record", but is less clear when we consider
   the user could hit "Record" and manually trigger page navigations.
   We'd have to pass a `didReload: true` type option through the code.

2. Change the tracing logic to copy that of Lighthouse (LH). LH will
   first navigate to `about:blank`, then trigger tracing, and then
   navigate to the user's target URL, thus ensuring that we start each
   recording from a completely blank trace. We can apply this logic only
   when the user hits "Reload & record", and no code that pertains to
   the rendering of the timeline needs any adjustment.


Given the above options, I went for (2) here. Not just because the code
changes are more contained, but we can have some confidence in this
approach given that it has been validated by LH.

Bug:1101268
Change-Id: I34f14c2f30bf3a91b234eebd6f032e8bff267b14