Shared Storage: Add DevTools metadata component for items view.
This adds the web component to display the metadata for an origin's
shared storage.

A later CL
(https://crrev.com/c/devtools/devtools-frontend/+/4006372)
will add a view that uses this component to show the metadata and
entries for a given origin's shared storage.

For the full initial Shared Storage DevTools UI prototype, see
https://crrev.com/c/devtools/devtools-frontend/+/3983292.

Bug:1372007,1218540
Change-Id: I011aef47e82bf0bf5f57472420f94effe1a4f01f