[sources] Do not automatically pretty print authored files.
Previously the "Automatically pretty print files" experiment would turn
on automatic pretty printing for all of JavaScript, CSS, and HTML,
regardless of whether that was authored or deployed code. In case of
sourcemapped files (which we consider authored), it definitely makes no
sense to automatically pretty print them (we still provide the option to
if the user decides this is useful).

Bug: chromium:1382752, chromium:1382755
Change-Id: I3133887bae578951d6d5f8455b70712d06904c10