[uma] Record metric for toggling to when user clicks on the group
Before we were tracking the toggling of expanding/collapsing
a breakpoint group via the toggle handler. However, toggling may also
happen when rendering the expanded group per default. In order to
capture whether the user is actually toggling, we are now moving
the tracking for whether the group was collapsed/expanded to the
click handler.

Bug: 1346231
Change-Id: Ibeae4e8df326113e365c2e1a37fd297b7e526d59