Add label to elements style filter
Adds a new label to the styles filter on the element page to indicate
how many style rules are hidden by a given filter.

The new label and the existing 'no matches' label are now both marked
as labels so screen readers will read them out as they change as a
result of the input, making the experience for those users much better.

There is an expected failure in blink web tests due to adding an ID to the filter bar. In order to land this change the following steps need to happen:
1) Disable the test (https://chromium-review.googlesource.com/c/chromium/src/+/4020775)
2) Land the change in devtools-frontend (This CL)
3) Update the test to expect the new ID and re-enable it (tracked by crbug.com/1383296)

Bug: 1379450
Change-Id: I56f215275759b7604b4e1471b9e8a3695951ea40