[sources] Resolve source URLs consistently
Currently, there is a subtle difference in relative path resolution for
source URLs in resource script mapping and in the source map manager: If
the base URL (the URL of the containing frame) cannot be used to make
the source URL absolute (for example, if the base URL is a data URL),
resource script mapping would interpret the URL as one without a domain,
but source map resolution would fail and the source map would not be
used.

With this patch, we change the URL resolution for source maps to be the
same as for source URL - if the path cannot be completed to an absolute
path, we just keep the URL as is.  The Pages pane will then interpret
the path as '(no domain)'.

Bug: chromium:1383416
Change-Id: If2f736535b0a09603f697e4be69864ca06d6ae24