[sources] Remove CodeMirror5 MIME type simplification logic.
In the CodeMirror5 days, we'd pass the MIME type to CodeMirror and let
it determine internally which highlighter to use, so we had to translate
between our own list of MIME types and what CodeMirror used. With the
current state of the codebase this logic is in CodeHighlighter.ts and
thus completely under our control, and simplifyMimeType() is thus
obsolete.

Also our handling of .php files was not correct, using the text/x-php
MIME type rather than the correct application/x-httpd-php MIME type[1].

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Common_types

Bug: chromium:1320204
Change-Id: Ibde231f759ccfa2d90007e30cee1adb12f3733df